ci: disable mac olink host tests
Olink host tests fail on mac. Those are neteork tests.
Most probably it is caused because either host or the sockets are not
ready to work and tests are already using it.
POCO does not provide easy mechanism to check if they are ready to use.
Untill proper wiaitng for them to be set up is introduced test are
disabled on macos.
For the bug see #142
This closes #148