feat(nats): handle request-response
Prepare the support to use request-response nats mechanism.
It will be necessary for handling methods remote calls.
The functionality added covers subbscribe, unsubscribe, storing callbacks,
Also disconnect added.
The implementation uses similar structures and functions as for subscriptions,
with details different - the callback added for topic has a return value.
In contrary to publish subscribe mechanism which is to be used for properties and signals,
only the requestor gets the response, not all the client adapters for same interface.
The request-response mechanism got its own thread pool in nats base.

This is part of #24 task.