feat(nats): make subscriptions non-blocking
Subscribe was a thread blocking call.
A thread pool was introduced to make those calls in a background.
The subscribe function also gets the subscribe callback which is
executed after the subscription action is finished either successfully
or not.
This will be used by the usesr to track the state of all its subscriptions.