feat: add publisher helper
Add a templated class as a publisher helper.
Its puropse is to help keep track of subscribers:
it allows subscribing and provide an id to unsubscribe, on publishChange
call it notifies all the subscribers about the change.
It is templated to allow different types of the changed value, and to
publish signals with many parameters.
It does not store any value.
It is thread safe.
The id generating mechanism may be improved.

It will be used in next commit to keep track of the nats adapters state.
It also may be used for generated interface publishers.

This is part of #24 task.