feat(nats): add base adapter
This commit introduces a base class for the generated adapters,
both client and service side.
The functionality it covers:
handling is ready for an adapter- where is ready means that the adapter
has working connection and has sucessfully subscribed to all topics it needed.
It also keeps track of the subscription state.

This is part of th #24 task.