feat(nats): client adapter implementation
Client Interface adapters now are using nats protocol.
They us base adatapter as a base class.
As the base requires weak_ptr it inherites enabl_shared_from_this.
The weak_ptr is necessary for making sure that callbacks passed to nats
client are always safe to call especially during tear down.
This is also why it has a create method instead of public constructor,
and it has an init method - to which all the logic for initialization
was moved to ensure no virtual method will be called on construction.

For the request response mechanism - if there is no responder the
returned value in an empty string, which causes crash on trying to
parse. This could be solved in a different way - detecting if there is a
service side running, waiting for init perhaps - but it won't work in
all cases as for now like when there was a service but it got
disconnected. If such functionality is required it need to be
implemented covering all connect-disconnect scenarios.

This is part of task #24.