feat(nats): nats service adapters
Service Interface adapters now are using nats protocol.
They us base adatapter as a base class.
As the base requires weak_ptr it inherites enabl_shared_from_this.
The weak_ptr is necessary for making sure that callbacks passed to nats
client are always safe to call especially during tear down.
This is also why it has a create method instead of public constructor,
and it has an init method - to which all the logic for initialization
was moved to ensure no virtual method will be called on construction.