feat(nats): add thread safety to base adapter
intorduces mutexes for stored subscriptions,
ensures that the callbacks won't be used after the object deleted
fixes checking for is ready:
in case the subscriptions are added one by one after the connection was
established the is connected callback was called after each subscription
added (the vector size didn't had its full size yet). Now it is ensured
that the isReady will actually wait for all the subscriptions.