feat(nats): support debug mode
The CNATS library has a d suffix for debug version.
As far as the cmake easy supports this with linking debug or optimised version, the conan has some issues with it.
For conan build there is the name of the library set in conan file, and later passed to cmake.

This is part of #24 task.