feat(nats): expose client Id
this is preparatory change for handiling cilent initialization with service side state
each client obtains the id on connection(valid through that connection
session).
This id will be used for initialization - the init message from service
is send to each client on demand (init request), and the id is in the
message topic.

This is part of the #24 task.