initialize nats client with service state:
each client subscribes for init message,
after successfull subscription of all topics it sends init request with its id in
payload, the topic contains the module and interface name.
If there is a service, already subscribed for such init request it
serves the message:
sends back one message wiht all the properties, only for this client
(the topic contains the id, and for such topic the client is
subscribed).
For the case when no service at time the client publishes the init request,
there is a messege with the interface service adapter availablity published
by the service once it has subscribed for all topics.
All the client interfaces adapters subscribe for this message.
This does not affect the is_ready functionality - which, as it was only
says the client is connected to NATS server with all its topic
subscribed, and not if it has connection with a service side of
interface.

This is part of #24 task.