initialize nats client with service state:
each client on subscribes for init message,
after successfull subscription it sends init request with its id in
payload, the topic contains the module and interface name.
If there is a service, already subscribed for such init request it
serves the message:
sends back one message wiht all the properties, only for this client
(the topic contains the id, and for such topic the client is
subscribed).
If there is no service at time the client publishes the init request,
the client will not reveive init(unless it will be re-connecter).
Therefor the service side after is connected sends message for each of
its property with the property value.
This does not affect the is_ready functionality - which, as it was only
says the client is connected to NATS server with all its topic
subscribed, and not if it has connection with a service side of
interface. Such functionality must be implemented if necessary.

This is part of #24 task.