feat(nats): handle subscriptions and messages
Adds handling subscriptions, sending and receiving messages.
The subscribe function has a callback for a topic to be handled.
Each subscription - even for the same topic, gets different id.
The ids are used to unsubscribe a callback from a topic.
The subscriptions and message handlers are stored in non-relocable containers.
This is very important, as the pointers to those are passed to NATS handlers.
Also a handler for errors for subscriptions is added during connection configuration.

The callbacks are stored before a subscription is made,
because a context with it is passed in form of a pointer to storage, to the subscribe function.

The subscriptions are wrapped with smart pointer (if successfully allocated).

Each subscription has also set a callback to clean it when it stops being used:
it is guaranteed that no message will be received on this subscription and it is ready to be cleaned.
This happens either after unsubscribe call or when connection gets closed.

This is part of #24 task.