feat(nats): add nats with empty impl
This is a first commit for implementing the nats protocol.

The CNATS libray is used for handling protocol.
The idea is to separate this library, from the code exposed to user, so it will be used only through the apigear/nats wrapper.
This commit introduces almost empty initialization. The functionality will be added with further commits.
The dependency for the CNATS library is added to conan feature.
It is possible to build it with CMake too, but you need to provide the library (you can use conan to get the dependencies only, see the CI workfolw for CMake build)

The requirement for the conan is added with visibility set to false - we wan't it to be build as static library and as a implementation detail, therefore it is not passed any further to user of the apigear/nats users.

This is part of #24 task.