feat(nats): handle connect
CWrapper now connects to a NATS server to given addres.
The connect function also allows adding a callback to be informed about connection status changes.

All the resources provided by th elibrary are wrapped with smart pointers with custom deleters, as the nats library is one that stores the requested items (connection, options, messages).
The class also stores a connection context, which is passed the CWrapper callbacks. The context contains the self weak_ptr, hence the

An option that is worth mentioning is natsOptions_UseGlobalMessageDelivery, without it set to true, the CNATS library would use separate thread for each subscription (topic). We assume to have a lot of subscriptions, that is why we decided to set it to true, which means, there is some thread pool which will be used instead.

Basic subscription handling was added just for test purposes. Proper implementation will be delivered in coming commits.
This is part of #24 task.