feat(nats): add flush
give better control over to user for sending messages.
The connection is configured in a wayt the messages are stored
for a period of time or untill buffer full.
With a flush function exposed, the user can force nats client to send
the messages. Nats offers also option to send right away each message
but it needs to be configured on connection start, and is same for all
the messages see the natsOptions_SetSendAsap. If that is desired
behavior then it should be used instead of flush which is less efficient
as it requires a round-trip with the server.

This is part of #24 task.