feat(olink): generated tests for olink
add generation tests for olink for the defined interfaces.
Tests do not use the network stack, the client and service are directly
connected (client send method with server receive method and vice
versa).
Tests are using the filter for non default value for singals and
properties.
The methods use default values and should be modified when customer
implementation added. Now only correctness of making a call or async
call is checked.

Ticket number #11