feat(nats): configure connection
Adds option to set sending each message as soon as possibe without
waiting for the buffer to be filled (or max time to pass).
As default the option is set to true as for local tests its latency and
test time even for lot of messages are way smaller.
Depending on your requirements for application you may want to set this
option to false.