feat(mqtt): increase safety for tear down client
C Mqtt client is now stored in unique_ptr with custom deleter.
Now when cwrapper is deleted at any point the CMqtt client will be
properly released.
There is also added a check if after 1000ms the request to do anything
is still valid (checking a weak_ptr)
This task is part of #3 task, which require often setup and teardown for
same client.