feat(mqtt): track subscription state
This commit introduces a base cass for interface adapters,
which purpose is to take care of subscriptions, its state and allow users of
the interface subscribe for events:
ready - which means the interface is connected and all subscriptions are
already made to mqtt broker
is unsubscribed - when all the subscriptions are removed from the
broker.
This is part of #3 task - adding generated mqtt tests.
The first event (read) was necessery to properly set up test, and start
it when the subscriptions are ready
The second was added for symmetry.