chores: clean up publisher
publisher needed lot of very similar members and behaviors for each
property and signal.
This could be gathered in a class for each property or signal.
Such class was already added and resides in apigear utilities.
This commit uses it with a small modification.
This is solves the #71 task from template-cpp14 repository (all tickets
from there are valid for cpp-17 but were not transfered).
It is solved only for cpp17 branch as this is enhancement, and for cpp14
only bugfixes are delivered.