fix(mqtt): clean up on disconnected
the on disconnected callbacks were moved to when the mqtt client states
it has disconnected (which is slightly later then disconnect function's
end.
The thread to reset the pointer and reconnect is starting now only to
when it really is needed. It used to be needed for all stituations to
destroy mqtt client pointer with paho functions, but now with the
wrapper is will be cleaned up anyway.